% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_postgres_db.R
\name{connect_to_postgres_db}
\alias{connect_to_postgres_db}
\title{Connect to a PostgreSQL db}
\usage{
connect_to_postgres_db(.dbname = NULL, .host = "localhost",
  .port = 5432, .user = NULL, .password = NULL)
}
\arguments{
\item{.dbname}{character name of dt}

\item{.host}{character host name, defaults to "localhost"}

\item{.port}{integer port number, defaults to 5432}

\item{.user}{character username}

\item{.password}{character password}
}
\value{
connection object
}
\description{
Connect to a PostgreSQL db
}
