% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_digits_to_words.R
\name{convert_digits_to_words}
\alias{convert_digits_to_words}
\title{Convert digits into English words}
\usage{
convert_digits_to_words(digits = NULL)
}
\arguments{
\item{digits}{character or numeric with digits to convert}
}
\value{
character with digits converted to words
}
\description{
Convert single digits (e.g. 1, 2, 3) into English words (one, two, three); does not yet support numbers with multiple digits
}
