% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_words_to_digits.R
\name{convert_words_to_digits}
\alias{convert_words_to_digits}
\title{Convert English words for numbers into digits}
\usage{
convert_words_to_digits(words = NULL, return_numeric = NULL)
}
\arguments{
\item{words}{words to turn into digits}

\item{return_numeric}{logical, set to true if output should be converted to numeric}
}
\value{
character unless \code{return_numeric == TRUE} then numeric
}
\description{
Convert English number words (e.g. one, two, three) into digits (1, 2, 3).
}
