% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_file_and_rename.R
\name{copy_file_and_rename}
\alias{copy_file_and_rename}
\title{Copy a file to a new directory and give it a new name}
\usage{
copy_file_and_rename(file_to_move = NULL, new_dir = NULL,
  new_name = NULL)
}
\arguments{
\item{file_to_move}{full path of file to copy}

\item{new_dir}{directory to copy file}

\item{new_name}{new name for file to be copied}
}
\value{
nothing
}
\description{
Copy a file to a new directory and give it a new name
}
