% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_newest_file.R
\name{identify_newest_file}
\alias{identify_newest_file}
\title{Identify newest file in a directory}
\usage{
identify_newest_file(dir = NULL, pattern = "\\\\.")
}
\arguments{
\item{dir}{character directory to search for files; should not end with slash}

\item{pattern}{character regex pattern for files}
}
\value{
character path to newest file
}
\description{
Identify newest file in a directory
}
