% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_char_to_numeric.R
\name{convert_char_to_numeric}
\alias{convert_char_to_numeric}
\title{Convert character vectors that could be numeric, to numeric}
\usage{
convert_char_to_numeric(char = NULL, threshold = 0.8)
}
\arguments{
\item{char}{character vector that could be numeric}

\item{threshold}{proportion of non-numeric values in char}
}
\value{
a numeric, or a character vector if threshold exceeded
}
\description{
Convert character vectors that could be numeric, to numeric
}
